requirements.txt
Django==3.2
numpy==1.19.2
pandas==1.1.3
IMDbPY==2021.4.18
requests==2.24.0
tmdbv3api==1.7.5